+++ pkg/database/bilidb.go
@@ -55,7 +55,7 @@ func BilGet(GroupID int64, MemberID int64, Status string) []LiveStream {
 		Limit = 2525
 	}
 
-	rows, err := DB.Query(`SELECT * FROM Vtuber.LiveBiliBili Inner join Vtuber.VtuberMember on VtuberMember.id=VtuberMember_id Inner join Vtuber.VtuberGroup on VtuberGroup.id = VtuberGroup_id Where (VtuberGroup.id=? or VtuberMember.id=?) AND Status=? Order by ScheduledStart DESC Limit ?`, GroupID, MemberID, Status, Limit)
+	rows, err := DB.Query(`SELECT LiveBiliBili.* FROM Vtuber.LiveBiliBili Inner join Vtuber.VtuberMember on VtuberMember.id=VtuberMember_id Inner join Vtuber.VtuberGroup on VtuberGroup.id = VtuberGroup_id Where (VtuberGroup.id=? or VtuberMember.id=?) AND Status=? Order by ScheduledStart ASC Limit ?`, GroupID, MemberID, Status, Limit)
 	if err != nil {
 		log.Error(err)
 	}